requirements.txt
certifi==2018.8.24
chardet==3.0.4
dbus-python==1.2.8
decorator==4.3.0
evento==1.0.1
idna==2.7
imageio==2.4.1
moviepy==0.2.3.5
numpy==1.15.2
omxplayer-wrapper==0.2.5
Pillow==5.3.0
requests==2.19.1
tqdm==4.26.0
urllib3==1.23
